+++ tests/TestCase.php
@@ -33,6 +33,7 @@ class TestCase extends Orchestra
     public function getEnvironmentSetUp($app): void
     {
         config()->set('database.default', 'testing');
+        config()->set('app.key', 'base64:'.base64_encode(str_repeat('a', 32)));
     }
 
     /**